requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
email-validator
sqlalchemy==2.0.23
psycopg2-binary
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.2
python-dotenv==1.0.0
alembic==1.12.1 
pydantic-settings>=2.0.0
python-slugify